Compute frames on which SmaxN has to be computed

This function computes a dataframe containing the minimal time needed for an individual of the studied species to go from a camera to another camera. It is based on the distance between each pair of cameras and the swim speed of the species.

compute.cam.time(dist_df, fish_speed)

Arguments

dist_df
a numerical dataframe containing the distance between each pair of camera. There are as many rows as there are cameras and there are as many columns as there are cameras, thus the dataframe is symmetrical and the diagonal is filled with 0. Rows names and columns names must be cameras names.

fish_speed
a numerical value refering to the maximal speed of the studied species. Speed must be given in meters per second.

Value

the function returns a dataframe containing the minimal time needed for an individual of the studied species to go from a camera to another camera.There are as many rows as there are cameras and there are as many columns as there are cameras, thus the dataframe is symmetrical and the diagonal is filled with 0.

@examples # Build distance dataframe for the example: dist_df_ex <- data.frame("A" = c(0, 2, 5, 5), "B" = c(2, 0, 5, 5), "C" = c(5, 5, 0, 4), "D" = c(5, 5, 4, 0)) rownames(dist_df_ex) <- c("A", "B", "C", "D")

# Run the function: compute.cam.time(dist_df = dist_df_ex, fish_speed = 1.6)