Tell if possible cells for the first cameras are possibly leading to the SmaxN or if they are not.

This function completes the first_cell_df which contains the cells for a given value and the first camera by adding TRUE or FALSE if the given cell can lead to the SmaxN or will not lead to the SmaxN by comparing to the maximal SmaxN computed on small UI.

first.cam.possible(time_df, first_cell_df, SmaxN_small_UI, bloc)

Arguments

time_df

a numerical dataframe containing the minimal time needed for an individual of the studied species to go from a camera to another camera.There are as many rows as there are cameras and there are as many columns as there are cameras, thus the dataframe is symmetrical and the diagonal is filled with 0. This dataframe is the output of the compute.cam.time function.

first_cell_df

a dataframe with value, cam_nm and timesteps of possible cells for the first camera

SmaxN_small_UI

a numeric value referring to the maximal value of the SmaxN for the small UI

bloc

the studied big UI

Value

the completed first_cell_df

Examples

 abund_df <- data.frame("A" = c(9,8,3,3,3,3,3), 
 "B" = c(0,4,2,2,1,3,3), 
 "C" = c(0,0,0,0,1,1,1), 
 "D" = c(1,0,0,0,3,3,3))
 
 time_df <- data.frame("A" = c(0,5,6,6), "B" = c(5,0,3,3), 
 "C" = c(6,3, 0, 3), "D" = c(6,3,3,0))
 rownames(time_df) <- c("A", "B", "C", "D")