Compute the Synchronised MaxN for a given second and a given camera

This function computes the Synchronised MaxN (SmaxN) for a given second and a given camera. It used the time dataframe gathering the time it takes for an individual of a given species to go from one camera to another.

compute.max.abund(dist_df, fish_speed, abund_df)

Arguments

dist_df
a numerical dataframe containing the distance between each pair of camera. There are as many rows as there are cameras and there are as many columns as there are cameras, thus the dataframe is symmetrical and the diagonal is filled with 0. Rows names and columns names must be cameras names.

fish_speed
a numerical value refering to the maximal speed of the studied species. Speed must be given in meters per second. If the computation of maxN values should not take into account fish speed (that is to say if the camera pooling is done at the second level), fish_speed = NULL

abund_df
a numerical dataframe containing the abundance of a given species across continuous time for several cameras. The columns refer to the cameras and the rows refers to the time. Time must be given in seconds and be continuous. BE CAREFUL that the cameras are in the same order in the abund_df and the time_df!.

Value

if the fish speed is taken into account, the function returns in a list: the Synchronised maxN (SmaxN) value (the maximal value computed by gathering temporally synchronised cameras) , the maxN value (the maximal value computed as the maximal abundance across all cameras and timesteps) , the maxN_cam values (a vector of maximal values for each camera, thus the list contains as many elements as there are cameras) , the maxN_row values (a vector of maximal values for each row by taking the maximal value of cameras abundances on each row, thus the list contains as many elements as there are rows in the abund_df dataframe) and the SmaxN_row value (the maximal value of the abundances summed across cameras on each row). If fish speed is not taken into account then, the returned list does not contain the SmaxN value because the SmaxN value is the equal to the SmaxN_row value.

Examples

 # Build distance dataframe for the example:
 dist_df_ex <- data.frame("A" = c(0, 2, 5, 5), "B" = c(2, 0, 5, 5), 
 "C" = c(5, 5, 0, 4), "D" = c(5, 5, 4, 0))
 rownames(dist_df_ex) <- c("A", "B", "C", "D")
 
 # Build distance dataframe for the example:
 abund_df_ex <- data.frame("A" = c(0, 1, 3, 7, 2, 2, 3, 0, 6, 2, 0, 1), 
                           "B" = c(2, 2, 2, 2, 0, 0, 0, 0, 1, 2, 1, 0), 
                           "C" = c(2, 0, 1, 0, 0, 4, 2, 2, 3, 0, 0, 4), 
                           "D" = c(0, 1, 0, 1, 0, 6, 1, 1, 6, 4, 2, 1))
 
 # Run the function:
 compute.max.abund(dist_df = dist_df_ex, 
                fish_speed = 1.6, 
                abund_df   = abund_df_ex)
#> $SmaxN
#> [1] 19
#> 
#> $maxN
#> [1] 7
#> 
#> $maxN_cam
#> A B C D 
#> 7 2 4 6 
#> 
#> $maxN_row
#>  [1] 2 2 3 7 2 6 3 2 6 4 2 4
#> 
#> $SmaxN_row
#> [1] 16
#>